Hybridize MSD sort with insertion sort
Notice that strings with long common prefixes, for example,

123456789012345567
aaaaaaaaaaaaabbbbbb
aaaaaaaaaaaaaaaabbbbbbbbbbbbbbcc
aaaaaaaaaaaaaaaabbbbbbbbbbbbbaccddd
aaaaaaaaaaaaaaaaaaaa

really hit the performance of MSD radix sort. In the example
above the first 13 letters of the 4 strings are all the same
and we will need 14 passes just to sort the first string out,
so does many more passes to sort the 2nd and the 3rd out.
In each pass (recursive call) we have to build a count array
of size R (usually 256), in order just to sort 4 strings!!

So, for such cases (lots of long common prefixes) applying
a hybrid with insertion sort can really boost performance.