Fix a bug in TST `find` routine
When we find the last TST node given the search string, the node
itself may not contain a value. For example, we've constructed
a TST tree that contains the string "thread-safe", and now if
we try to search for the string "thread", a valid TST node will
be found. But the node itself isn't associated with a value.